fix Clear link in Gunning Fog Index module
Module was  initially developed and tested with AngularJS 1.3.8. After merging with master AngularJS version was bumped to 1.3.16. In new version instead of looking up non-prefixed scope member in parent scope, a new member was created in directive's scope. After placing "text" to controller scope and referring it with "fog.text" correct scoping is applied and issue is resolved.

Thanks to @D049677 for finding it in #885.

close #889