[trafficLightService] status Off priority
Commit 691911634dbfcb3aa46a588af301fa9f215cce80 added support
for the status Off, but in the way it was implemented, for the
MultiApp support, status off was taking priority over status
Green, which is wrong. If one app shows Green and the others
are Off, than it should show green.

This patch fixes it and introduces Automated tests for the
service.