[trafficLightService] Refactoring
Some redundancies were removed and there was an unifaction of the term
used for the trafficLight state.

Before this patch both the terms state and color were used.
Since the state Off have been introduced, the term state makes more sense
since you can think of Green, Red and Yellow as states, but is does not
fit so well to think of Off as color.